requirements.txt
python-dotenv
pydantic==2.12.4
langchain==1.1.3
langchain-core==1.1.3
langchain-google-vertexai
langchain-text-splitters
langchain-community
langchain-mcp-adapters
langchain-anthropic
langchain-openai
langgraph
deepagents==0.3.0
rich==14.2.0
requests
pyserial==3.5
readchar==4.2.1
opencv-python
opencv-python-headless